+++ Journal-Services/Infrastructure/Journal.Infrastructure/Features/Driver/IDriverRepository.cs
@@ -17,5 +17,5 @@ public interface IDriverRepository
     /// </summary>
     /// <param name="dto">The driver to be saved</param>
     /// <returns>The ID of the driver.</returns>
-    int InsertDriver(DriverDTO dto);
+    long InsertDriver(DriverDTO dto);
 }
